pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "beautifulsoup4>=4.12.0",
    "wikipedia>=1.4.0",
    "duckduckgo-search>=4.0.0",
    "tavily-python>=0.3.0",
]
